+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 
+- -p flag to texter for showing path to file
+- Simplified markdown and text rendering
 - 'rm' command to remove tagged objects
 - Tag events for post modification
 - texter example for rendering text files